+++ CHANGELOG.md
@@ -4,6 +4,12 @@
 
 ## [Unreleased] <!-- ReleaseDate -->
 
+### Features
+
+- The `CustomType` and `Confirm` prompts now support auto-submitting on valid input. [#321](https://github.com/mikaelmello/inquire/pull/321).
+
+### Dependencies
+
 - Bumped MSRV from 1.80 -> 1.82 due to new requirements of dependencies.
 
 ## [0.9.1] - 2025-09-16